trying something

+++ sobel-gpu-bmtalotta-cs441.cu
@@ -50,7 +50,7 @@ __global__ void sobel(int width, int height, char *pixels, int *arr)
     y22 *= pixels[pixelIndex(x+1,y+1,width)];
     int pixelX = x00 + x01 + x02 + x20 + x21 + x22;
     int pixelY = y00 + y10 + y20 + y02 + y12 + y22;
-    arr[pixelIndex(x,y,width)] = (int)sqrt(pixelX*pixelX +pixelY*pixelY);
+    arr[pixelIndex(x,y,width)] = (int)sqrtf(pixelX*pixelX +pixelY*pixelY);
   }else{
     arr[pixelIndex(x,y,width)] = pixels[pixelIndex(x,y,width)];
   }
